First commit
Update SimpleAnimationStarter to draw a frame with a gray background and a strokeRect. A lot of effort getting Netbeans and Maven to work with JavaFX, understanding pom.xml and updating environ vars, plus getting latest maven binary + setting path so I could see maven in the command line and use it to run the project. Too much effort!