Finish chapter 3, exercise 9
Updated SimpleAnimationStart so we draw some squares that oscillate back and forth, left to right on the screen.  We use the width of the frame and the frame number to determine when the animation can start to move in the opposite direction. Added a for loop to draw multiple squares where the y co-ord is the size of the length of the square multiplied by the value of the value incremented in the loop.

Can adjust the speed by setting a speed variable and multiplying all occurrences of the frame number by the speed. Can have different speeds for each box by updating the speed variable with the value being incremented in the for loop.